+++ shuup/xtheme/admin_module/views/_snippet.py
@@ -19,7 +19,7 @@ from shuup.admin.utils.picotable import Column, TextFilter
 from shuup.admin.utils.views import CreateOrUpdateView, PicotableListView
 from shuup.apps.provides import get_provide_objects
 from shuup.core import cache
-from shuup.xtheme.admin_module.widgets import CodeEditorWidget
+from shuup.xtheme.admin_module.widgets import XthemeCodeEditorWidget
 from shuup.xtheme.models import Snippet
 from shuup.xtheme.resources import GLOBAL_SNIPPETS_CACHE_KEY
 
@@ -29,7 +29,7 @@ class SnippetForm(forms.ModelForm):
         model = Snippet
         fields = ["location", "themes", "snippet_type", "snippet"]
         widgets = {
-            "snippet": CodeEditorWidget()
+            "snippet": XthemeCodeEditorWidget()
         }
 
     def __init__(self, *args, **kwargs):